fix: [Data history]: Fixed data grouping for large data set (#742)
* fixed import for data grouping to work

* moved mock data to mock file

* lint fix
